Time: 0 ms (100.00%) | Memory: 22 MB (47.88%) - LeetSync

+++ LeetcodeSolutions/98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -1,7 +1,7 @@
 
 class Solution {
 private:
-    bool helper(TreeNode* root,long long int lRange,long long int RRange){
+    bool helper(TreeNode* root,long  lRange,long  RRange){
         if (!root) return true;
         if(!(root->val > lRange && root->val<RRange)) return false;
         return helper(root->left,lRange,root->val) && helper(root->right,root->val,RRange);
@@ -10,4 +10,4 @@ public:
     bool isValidBST(TreeNode* root) {
         return helper(root,(1L*INT_MIN)-1,(1L*INT_MAX)+1);
     }
-};
+};
